requirements.txt
cohere
openai
python-dotenv
streamlit